+++ core/src/main/java/com/graphhopper/util/details/PathDetailsBuilderFactory.java
@@ -44,6 +44,11 @@ public class PathDetailsBuilderFactory {
         if (requestedPathDetails.contains(LEG_WEIGHT))
             builders.add(new ConstantDetailsBuilder(LEG_WEIGHT, path.getWeight()));
 
+        for (String key : requestedPathDetails) {
+            if (key.endsWith("_conditional"))
+                builders.add(new KVStringDetails(key));
+        }
+
         if (requestedPathDetails.contains(STREET_NAME))
             builders.add(new KVStringDetails(STREET_NAME));
         if (requestedPathDetails.contains(STREET_REF))
@@ -73,7 +78,8 @@ public class PathDetailsBuilderFactory {
             builders.add(new IntersectionDetails(graph, weighting));
 
         for (String pathDetail : requestedPathDetails) {
-            if (!evl.hasEncodedValue(pathDetail)) continue; // path details like "time" won't be found
+            if (!evl.hasEncodedValue(pathDetail))
+                continue; // path details like "time" won't be found
 
             EncodedValue ev = evl.getEncodedValue(pathDetail, EncodedValue.class);
             if (ev instanceof DecimalEncodedValue)
@@ -86,7 +92,8 @@ public class PathDetailsBuilderFactory {
                 builders.add(new StringDetails(pathDetail, (StringEncodedValue) ev));
             else if (ev instanceof IntEncodedValue)
                 builders.add(new IntDetails(pathDetail, (IntEncodedValue) ev));
-            else throw new IllegalArgumentException("unknown EncodedValue class " + ev.getClass().getName());
+            else
+                throw new IllegalArgumentException("unknown EncodedValue class " + ev.getClass().getName());
         }
 
         if (requestedPathDetails.size() > builders.size()) {